requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.1
attrs==25.3.0
boto3==1.37.21
botocore==1.37.21
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==7.5.4
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
griffe==1.6.3
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jiter==0.9.0
jmespath==1.0.1
logfire-api==3.11.0
lxml==5.3.1
markdown-it-py==3.0.0
mcp==1.5.0
mdurl==0.1.2
mistralai==1.6.0
multidict==6.2.0
ollama==0.4.7
openai==1.68.2
opentelemetry-api==1.31.1
packaging==24.2
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai==0.0.46
pydantic-ai-slim==0.0.46
pydantic-graph==0.0.46
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-inspection==0.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
youtube-transcript-api==1.0.3
zipp==3.21.0
